# todo-app/user.go
package todo_app

type User struct {
	Id       int    `json:"-" db:"id"`
	Name     string `json:"name" binding:"required"'`
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required"`
}